# biz-case/spring-boot-spec-demo-ai/backend/src/main/java/com/example/openspecdemo/dto/auth/LoginRequest.java
package com.example.openspecdemo.dto.auth;

import jakarta.validation.constraints.NotBlank;

public class LoginRequest {

    @NotBlank
    private String username;

    @NotBlank
    private String password;

    public String getUsername() {
        return username;
    }

    public void setUsername(String username) {
        this.username = username;
    }

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }
}