# biz-case/spring-boot-spec-demo-ai/backend/src/main/java/com/example/openspecdemo/dto/user/UserResponse.java
package com.example.openspecdemo.dto.user;

public class UserResponse {

    private Long id;

    private String username;

    private String email;

    public UserResponse() {
    }

    public UserResponse(Long id, String username, String email) {
        this.id = id;
        this.username = username;
        this.email = email;
    }

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public String getUsername() {
        return username;
    }

    public void setUsername(String username) {
        this.username = username;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }
}