# biz-case/spring-boot-spec-demo-ai/backend/src/main/java/com/example/openspecdemo/dto/user/UserUpdateRequest.java
package com.example.openspecdemo.dto.user;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;

public class UserUpdateRequest {

    @NotBlank
    @Email
    private String email;

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }
}